+++ src/main/java/com/sportsify/ticketing/domain/model/OrderStatus.java
@@ -1,5 +1,5 @@
 package com.sportsify.ticketing.domain.model;
 
 public enum OrderStatus {
-    PENDING, CONFIRMED, CANCELLED
+    PENDING, PAYING, CONFIRMED, CANCELLED, EXPIRED
 }
