+++ src/main/java/com/sportsify/ticketing/domain/repository/OrderRepository.java
@@ -1,7 +1,13 @@
 package com.sportsify.ticketing.domain.repository;
 
 import com.sportsify.ticketing.domain.model.Order;
+import com.sportsify.ticketing.domain.model.OrderStatus;
+
+import java.time.LocalDateTime;
+import java.util.List;
 
 public interface OrderRepository {
     void save(Order createdOrder);
+
+    List<Order> findByStatusAndExpiresAtBefore(OrderStatus status, LocalDateTime expiresAt);
 }
